stopped touching all memory.  started adding syncrecord unquantized

+++ src/gui/slider_bar.cpp
@@ -33,8 +33,8 @@ using namespace std;
 
 // Convert a value in dB's to a coefficent
 #undef DB_CO
-#define DB_CO(g) ((g) > -144.0f ? powf(10.0f, (g) * 0.05f) : 0.0f)
-#define CO_DB(v) (20.0f * log10f(v))
+#define DB_CO(g) ((g) > -144.0f ? pow(10.0f, (g) * 0.05f) : 0.0f)
+#define CO_DB(v) (20.0f * log10(v))
 
 static inline double 
 gain_to_slider_position (double g)
